gpu stuff

+++ dev/SupportingFunctions.py
@@ -49,7 +49,7 @@ def hard_threshold_k(X, k):
 	m = X.data.shape[1]
 	a,_ = torch.abs(Gamma).data.sort(dim=1,descending=True)
 	T = torch.mm(a[:,k].unsqueeze(1),torch.Tensor(np.ones((1,m))))
-	mask = Variable(torch.Tensor( (np.abs(Gamma.data.numpy())>T.cpu().numpy()) + 0.))
+	mask = Variable(torch.Tensor( (np.abs(Gamma.data.numpy())>T.cpu().numpy()) + 0.)).to(dtype=dtype)
 	Gamma = Gamma * mask
 	Gamma = Gamma.view(X.data.shape)
 	return Gamma, mask.data.nonzero()
@@ -70,7 +70,7 @@ def sample_filters(numb_atoms, p, k):
 	return active_filter_inds
 
 def create_dropout_mask(numb_dp, numb_atoms, numb_r, numb_c, active_filter_inds):
-	temp = torch.zeros(numb_atoms, numb_r, numb_c)
+	temp = torch.zeros(numb_atoms, numb_r, numb_c).to(dtype=dtype)
 	for i in active_filter_inds:
 		temp[i] = torch.ones(numb_r, numb_c)
